+++ src/Azure/Storage.Net.Microsoft.Azure.Storage.Files/Files.csproj
@@ -1,0 +1,25 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Storage.Net.Microsoft.Azure.Storage.Files</AssemblyName>
+    <RootNamespace>Storage.Net.Microsoft.Azure.Storage.Files</RootNamespace>
+     <LangVersion>latest</LangVersion>
+     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+     <WarningsAsErrors />
+  </PropertyGroup>
+
+   <PropertyGroup Condition="'$(Configuration)' == 'Release'">
+      <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+   </PropertyGroup>
+
+
+   <ItemGroup>
+      <PackageReference Include="Microsoft.Azure.Storage.File" Version="11.1.0" />
+   </ItemGroup>
+
+   <ItemGroup>
+     <ProjectReference Include="..\..\Storage.Net\Storage.Net.csproj" />
+   </ItemGroup>
+
+</Project>
