+++ src/storage.sln
@@ -55,6 +55,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Databricks.Dbfs", "Azure\St
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blobs", "Azure\Storage.Net.Microsoft.Azure.Storage.Blobs\Blobs.csproj", "{6F4E9CCC-DEB7-4D40-B2BA-4914171D26B2}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Files", "Azure\Storage.Net.Microsoft.Azure.Storage.Files\Files.csproj", "{598A07EE-E6D5-4B54-81FB-1C0639847B18}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -121,6 +123,10 @@ Global
 		{6F4E9CCC-DEB7-4D40-B2BA-4914171D26B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6F4E9CCC-DEB7-4D40-B2BA-4914171D26B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6F4E9CCC-DEB7-4D40-B2BA-4914171D26B2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -145,6 +151,7 @@ Global
 		{A79A5CE0-83F3-44E2-A88D-D51D7A3EE8F5} = {7060A53B-F046-4820-A143-1A05B8C14575}
 		{8EED860F-4D6C-4EAE-97D3-62925E7336E0} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
 		{6F4E9CCC-DEB7-4D40-B2BA-4914171D26B2} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
+		{598A07EE-E6D5-4B54-81FB-1C0639847B18} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C5FC6E9A-255E-4FD4-B08C-8EDE44BFE37A}
