+++ test/Storage.Net.Tests.Integration/Storage.Net.Tests.Integration.csproj
@@ -28,6 +28,7 @@
     <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.KeyVault\KeyVault.csproj" />
     <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.ServiceBus\ServiceBus.csproj" />
     <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.Storage.Blobs\Blobs.csproj" />
+    <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.Storage.Files\Files.csproj" />
     <ProjectReference Include="..\..\src\Azure\Storage.Net.Microsoft.Azure.Storage\LegacyStorage.csproj" />
     <ProjectReference Include="..\..\src\GCP\Storage.Net.Gcp.CloudStorage\Storage.Net.Gcp.CloudStorage.csproj" />
     <ProjectReference Include="..\..\src\General\Storage.Net.Ftp\Storage.Net.Ftp.csproj" />
